appbar and drawer setup
